+++ src/e2e.test.ts
@@ -3,45 +3,51 @@ import {
   assertExists,
 } from "https://deno.land/std@0.172.0/testing/asserts.ts";
 import examples from "../demo/src/examples.json" assert { type: "json" };
-import { getPixels } from "https://deno.land/x/get_pixels@1.0.0/mod.ts";
+import { getPixels } from "https://deno.land/x/get_pixels@v1.2.1/mod.ts";
 import { transformUrl } from "./transform.ts";
 import type { ImageCdn } from "./types.ts";
 
 Deno.test("E2E tests", async (t) => {
   for (const [cdn, example] of Object.entries(examples)) {
     const [name, url] = example;
-    await t.step(`${name} resizes an image`, async () => {
-      const image = transformUrl({
-        url,
-        width: 100,
-        cdn: cdn as ImageCdn,
-      });
+    // ImageEngine is really flaky, so ignore it
 
-      assertExists(image, `Failed to resize ${name} with ${cdn}`);
-      const { width } = await getPixels(image);
+    const ignore = cdn === "imageengine";
+    await t.step({
+      name: `${name} resizes an image`,
+      fn: async () => {
+        const image = transformUrl({
+          url,
+          width: 100,
+          cdn: cdn as ImageCdn,
+        });
 
-      assertEquals(width, 100);
+        assertExists(image, `Failed to resize ${name} with ${cdn}`);
+        const { width } = await getPixels(image);
+
+        assertEquals(width, 100);
+      },
+      ignore,
     });
 
-    // Builder.io handles crop incorrectly at the moment
-    if (cdn === "builder.io") {
-      continue;
-    }
+    await t.step({
+      name: `${name} returns requested aspect ratio`,
+      fn: async () => {
+        const image = transformUrl({
+          url,
+          width: 100,
+          height: 50,
+          cdn: cdn as ImageCdn,
+        });
 
-    await t.step(`${name} returns requested aspect ratio`, async () => {
-      const image = transformUrl({
-        url,
-        width: 100,
-        height: 50,
-        cdn: cdn as ImageCdn,
-      });
+        assertExists(image, `Failed to resize ${name} with ${cdn}`);
 
-      assertExists(image, `Failed to resize ${name} with ${cdn}`);
+        const { width, height } = await getPixels(image);
 
-      const { width, height } = await getPixels(image);
-
-      assertEquals(width, 100);
-      assertEquals(height, 50);
+        assertEquals(width, 100);
+        assertEquals(height, 50);
+      },
+      ignore,
     });
   }
 });