+++ src/parse.ts
@@ -15,6 +15,9 @@ import { parse as scene7 } from "./transformers/scene7.ts";
 import { parse as keycdn } from "./transformers/keycdn.ts";
 import { parse as directus } from "./transformers/directus.ts";
 import { parse as imageengine } from "./transformers/imageengine.ts";
+import { parse as contentstack } from "./transformers/contentstack.ts";
+import { parse as cloudflareImages } from "./transformers/cloudflareimages.ts";
+import { parse as ipx } from "./transformers/ipx.ts";
 import { ImageCdn, ParsedUrl, SupportedImageCdn, UrlParser } from "./types.ts";
 
 export const parsers = {
@@ -33,7 +36,10 @@ export const parsers = {
   scene7,
   keycdn,
   directus,
-  imageengine
+  imageengine,
+  contentstack,
+  "cloudflare_images": cloudflareImages,
+  ipx,
 };
 
 export const cdnIsSupportedForParse = (