+++ src/types.ts
@@ -35,7 +35,7 @@ export interface ShouldDelegateUrl {
 }
 
 export type CdnOptions = {
-  [key in ImageCdn]: Record<string, unknown>;
+  [key in ImageCdn]?: Record<string, unknown>;
 };
 export interface UrlGeneratorOptions<TParams = Record<string, string>> {
   base: string | URL;
@@ -91,6 +91,9 @@ export type ImageCdn =
   | "scene7"
   | "keycdn"
   | "directus"
-  | "imageengine";
+  | "imageengine"
+  | "contentstack"
+  | "cloudflare_images"
+  | "ipx";
 
 export type SupportedImageCdn = ImageCdn;